Added some placeholder code for parallax effect, buttons and some text
Created a placeholder div to store the code for the parallax scrolling effect for later use. Added buttons that for now do not work and added some visuals to them. Finaly Added some text to represent the content that is to be displayed on that section, but not yet stylized.