Visually updated the hero section of the site and added a new placeholder section for more content
Somewhat finalized the landing screen of the site and created a card in the center of the image. Differentiated the two buttons in two categories, Primary and Secondary, in order to draw the attention of the user to certain pages. Created a second section that will be designed to be placed in between image sections, but not yet stylized.